+++ CHANGELOG.md
@@ -6,7 +6,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## 1.1.0
-Extend with possibility of requesting and receiving Key Attestations
+
+### Added
+- Enable requesting Key Attestation via Issuer Metadata
+- Enable receiving and verification of Key Attestations in Credential Request Proofs.
+
+### Fixed
+
+- Checks for protected claims are now done in the create-offer-flow (1 step) instead of the issuance flow.
+- Business Issuer is directly informed when the payload cannot be processed later.
+
+## 1.0.1
+
+### Fixed
+
+- Fix status code when jwt filter criteria are not met from a 500 to 401.
 
 ## 1.0.0
 Merge of issuer-agent-management 1.6.1 and issuer-agent-oid4vci into one service.